get rid of local client entirely
