+++ rpc/mod.ts
@@ -1,6 +1,5 @@
 export * from "./Base.ts";
 export * from "./messages.ts";
 export * from "./providers/detect.ts";
-export * from "./providers/local.ts";
 export * from "./providers/proxy.ts";
 export * from "./providers/smoldot.ts";
