+++ test-util/Chain.ts
@@ -1,19 +1,13 @@
-import { Beacon } from "../Beacon.ts";
 import { Chain } from "../core/Chain.ts";
 import { KnownRpcMethods } from "../known/mod.ts";
-import { LocalBeacon, LocalClientProps } from "../rpc/mod.ts";
+import { ProxyBeacon } from "../rpc/mod.ts";
+import { SubstrateProcess } from "../util/mod.ts";
 import { TestAddresses } from "./Addresses.ts";
 
-export class TestChain extends Chain<Beacon<LocalClientProps, KnownRpcMethods>> {
+export class TestChain extends Chain<ProxyBeacon<KnownRpcMethods>> {
   override address: TestAddresses<this> = new TestAddresses(this);
 }
 
-export function chain() {
-  return new TestChain(
-    new LocalBeacon({
-      path: "./node-template",
-      cwd: new URL(".", import.meta.url).pathname,
-      dev: true,
-    }),
-  );
+export function chain(process: SubstrateProcess) {
+  return new TestChain(new ProxyBeacon(process.url));
 }
