+++ util/mod.ts
@@ -3,4 +3,5 @@ export * from "./discovery_value_validation.ts";
 export * from "./env.ts";
 export * from "./ErrorCtor.ts";
 export * as hex from "./hex.ts";
+export * from "./substrateProcess.ts";
 export * from "./types.ts";
